requirements.txt
requests==2.22.0
pymongo[srv]==3.10.1
python-dotenv==0.14.0
lxml==4.6